requirements.txt
python-dotenv==1.0.1
google-generativeai==0.7.2
youtube-transcript-api==0.6.2
requests==2.32.3
beautifulsoup4==4.12.3